nvmetcli: remove status method for the UIPortNode
nvmetcli defines a status method for the UIPortNode class but there is
no corresponding 'enable' file in configfs so typing in status will
always result in nvmetcli printing disbaled if one types 'status' in
the port node:

/> cd ports/1/
/ports/1> status
Status for /ports/1: disabled
/ports/1>

There are two possible ways for fixing this issue, provide a configfs
file to enable/disable the port node in the kernel or delete the
method in nvmetcli. Having a configfs file in the kernel is pointless,
as ports get enabled when symlinked to a namespace so zap the pointless status
method in user-space.

Signed-off-by: Johannes Thumshirn <jthumshirn@suse.de>
Reported-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Christoph Hellwig <hch@lst.de>